Split widgets: action pill and status badge
