+++ extension/devtools/app/lib/main.dart
@@ -19,6 +19,8 @@ part 'hooks/use_signals_panel_state.dart';
 part 'hooks/use_timeline_panel_state.dart';
 part 'hooks/use_ui_state.dart';
 part 'utils/helpers.dart';
+part 'widgets/action_pill.dart';
+part 'widgets/status_badge.dart';
 
 void main() {
   runApp(const DevToolsExtension(child: OrefDevToolsApp()));
@@ -3174,36 +3176,6 @@ class _SignalDetail extends StatelessWidget {
             const SizedBox(height: 12),
             Text(entry!.note, style: textTheme.bodySmall),
           ],
-        ],
-      ),
-    );
-  }
-}
-
-class _StatusBadge extends StatelessWidget {
-  const _StatusBadge({required this.status});
-
-  final String status;
-
-  @override
-  Widget build(BuildContext context) {
-    final style = _statusStyles[status] ?? _statusStyles['Active']!;
-    return _GlassPill(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
-      color: style.color.withValues(alpha: 0.2),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Container(
-            width: 8,
-            height: 8,
-            decoration: BoxDecoration(
-              color: style.color,
-              shape: BoxShape.circle,
-            ),
-          ),
-          const SizedBox(width: 6),
-          Text(status),
         ],
       ),
     );
